test: Add test for large repo graph drawing
The repo, namely it's the micropython github repo with pull requests
fetched as heads, contains a lot of criss-cross merges and is a nice
candidate for acting as the base of a test.

The expected output is the current state of the graph drawing code.
We're going to update the output as we fix issues.

Signed-off-by: Sven Wegener <sven.wegener@stealer.net>